pyproject.toml
[project]
name = "polymarket-trading-bot"
version = "0.1.0"
description = "Telegram bot for Polymarket trading"
authors = [
    {name = "mobneo", email = "mobneo@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "aiogram (>=3.28.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "py-clob-client-v2 (>=1.0.0,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "bot" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
